requirements.txt
tensorflow
tensorflow-datasets
spacy
en_core_web_sm
opencv-python
tkinter
psutil
beautifulsoup4
requests
pytesseract
Pillow
